Made some progress on the transaction tag editor and the create account screen.

+++ src/Asm.BankPlus.Web/ClientApp/src/pages/settings/transactionTags/TransactionTagRow.tsx
@@ -45,7 +45,10 @@ function useTransactionTagRowEvents(props: TransactionTagRowProps) {
     }
 
     const deleteTag = () => {
-        dispatch(actionCreators.deleteTag(props.tag));
+
+        if (window.confirm("Deleting this tag will remove it from all rules and transactions. Are you sure?")) {
+            dispatch(actionCreators.deleteTag(props.tag));
+        }
     }
 
     const addTag = (tag: TransactionTag) => {
